requirements.txt
discord.py
youtube-dl
yt-dlp
spotipy
ffmpeg
python-dotenv
